+++ src/actividades/Quiz.java
@@ -10,13 +10,13 @@ import user.Profesor;
 
 public class Quiz extends Actividad {
 	private static String TIPO = "Quiz";
-    private float calificacionMinima;
+    private double calificacionMinima;
     private List<PreguntaCerrada> preguntas;
     private int numPreguntas;
 
     // Constructor
     public Quiz(String titulo, String descripcion, List<String> objetivos, String nivelDificultad, int duracionMinutos,
-    		Date fechaLimite, Profesor profesorCreador, float calificacionMinima, List<PreguntaCerrada> preguntas) 
+    		Date fechaLimite, Profesor profesorCreador, double calificacionMinima, List<PreguntaCerrada> preguntas) 
     
     {
         super(titulo, descripcion, objetivos, nivelDificultad, duracionMinutos, fechaLimite, profesorCreador);
@@ -34,11 +34,11 @@ public class Quiz extends Actividad {
 		return numPreguntas;
 	}
 
-	public float getCalificacionMinima() {
+	public double getCalificacionMinima() {
         return calificacionMinima;
     }
 
-    public void setCalificacionMinima(float calificacionMinima) {
+    public void setCalificacionMinima(double calificacionMinima) {
         this.calificacionMinima = calificacionMinima;
     }
